signage_requirements_Version2.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
gunicorn==21.2.0
Werkzeug==3.0.2
itsdangerous==2.2.0
bcrypt==4.1.3
passlib[bcrypt]==1.7.4
pillow==10.2.0
pyyaml==6.0.1
websockets==12.0
flask-sock==0.7.0
APScheduler==3.10.4
python-mpv==1.0.5
watchdog==4.0.0
requests==2.31.0